[exporter] Change specification to treat lilToon as glTF PBR under specific conditions
When lilToon's gloss type in gloss settings is "Real" and composite mode is the default "Additive", treat as glTF PBR instead of MToon. This makes it PBR-oriented and improves reproducibility, but may affect lighting as it's often used with matcap. However, enabling matcap in this case can cause over-lighting, so this change is adopted as a trade-off.

When lilToon's gloss type is "Real", BRDF (GGX + Schlick Fresnel) is used for specular reflection calculation. The basic calculation formula matches glTF, but diffuse reflection differs - glTF uses Lambertian while lilToon uses toon ramp. The final color output is the sum of diffuse and specular reflection, so they will not be identical.

Split source code into class-level files to facilitate the above changes. The previously monolithic source code is being split into per-class files to make implementation of the above changes easier.
