+++ Editor/Exporter.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea365e5d9b784a28a31c95b362381e8d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
