+++ Editor/Exporter/GltfMaterialExporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfa22d08bf6744bcafd03e1e52cc619e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
