+++ Editor/Exporter/NdmfVrmExporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1811dc6e86b4330868b91cb7de6a97b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
