+++ Editor/Exporter/VrmRootExporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2d3fdd0a5bb4eb683eb0668b24ac68f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
